Split shared/adminSettings into types + defaults (390 → 16 line barrel)
The file mixed 16 interfaces (~245 lines) with 4 large default
constants (~145 lines). Readers fixing a default never need the
interfaces; readers tracing a type rarely need the literal values.
Two-file split, barrel re-exports the public API so all 14
consumers keep working without changes.

- shared/admin-settings/types.ts      every interface — app prefs,
                                      personalization, managed users,
                                      auth, agent definitions, all
                                      per-integration shapes, and
                                      the AdminSettings composite
- shared/admin-settings/defaults.ts   defaultAppSettings,
                                      defaultPersonalizationSettings,
                                      defaultAgentDefinitions,
                                      defaultIntegrations with a
                                      head comment explaining the
                                      "secrets default to empty"
                                      policy
- shared/adminSettings.ts             16-line barrel — `export *
                                      from "./admin-settings/types"`
                                      + `./admin-settings/defaults"`

No behavior change — the re-exports preserve every previous import
path. Total LOC is up slightly (441 vs 390) because each file
carries its own import header; clarity is the win.
