+++ shared/admin-settings/defaults.ts
@@ -1,0 +1,167 @@
+/**
+ * Default values for the admin-settings shapes. These are the
+ * starting state when no admin-settings.json exists on disk, and
+ * the per-section reset in `resetAppSettings`. Edit here when you
+ * want to change "what shows up first-run".
+ *
+ * Secret-bearing fields (tokens, passwords, API keys) intentionally
+ * default to empty — they get populated via the admin UI, never
+ * shipped with the repo.
+ */
+
+import type {
+  AgentDefinition,
+  AppSettings,
+  IntegrationsSettings,
+  PersonalizationSettings,
+} from "./types";
+
+export const defaultAppSettings: AppSettings = {
+  notifications: true,
+  agentAlerts: true,
+  messageNotifications: true,
+  systemAlerts: true,
+  hapticFeedback: true,
+  autoSpellCorrect: true,
+  autoSendDictation: false,
+  backgroundConversations: true,
+  autocomplete: false,
+  trendingSearches: true,
+  followUpSuggestions: false,
+  colorScheme: "dark",
+  language: "English",
+  voiceType: "Ember",
+};
+
+export const defaultPersonalizationSettings: PersonalizationSettings = {
+  displayName: "Admin",
+  preferredLanguage: "English",
+  colorScheme: "dark",
+  compactMessages: false,
+  showTimestamps: true,
+  fontSize: "medium",
+};
+
+export const defaultAgentDefinitions: AgentDefinition[] = [
+  {
+    key: "OperationsAgent",
+    label: "Operations Agent",
+    status: "active",
+    description: "Handles operational tasks, approvals, and execution-oriented requests.",
+    entryPoint: "server/agents/operations/OperationsAgent.ts",
+  },
+  {
+    key: "IntelligenceAgent",
+    label: "R&D Agent",
+    status: "active",
+    description:
+      "Researches, synthesizes, and summarizes market, crypto, stock, and knowledge requests.",
+    entryPoint: "server/agents/intelligence/IntelligenceAgent.ts",
+  },
+  {
+    key: "BusinessManagerAgent",
+    label: "Business Manager Agent",
+    status: "active",
+    description:
+      "Business operations: payroll, contractors, ecommerce, real estate, business credit.",
+    integration: "Gusto",
+    entryPoint: "server/agents/business-manager/BusinessManagerAgent.ts",
+  },
+  {
+    key: "FinanceAgent",
+    label: "Finance Agent",
+    status: "active",
+    description: "Crypto, forex, trading setups, wealth strategy. Action-gated by approval.",
+    entryPoint: "server/agents/finance/FinanceAgent.ts",
+  },
+];
+
+export const defaultIntegrations: IntegrationsSettings = {
+  gusto: {
+    enabled: false,
+    status: "planned",
+    environment: "sandbox",
+    companyId: "",
+    apiBaseUrl: "https://api.gusto-demo.com",
+    clientId: "",
+    webhookBaseUrl: "",
+    notes: "Planned integration for payroll, contractors, onboarding, and benefits workflows.",
+  },
+  github: {
+    enabled: false,
+    status: "planned",
+    apiBaseUrl: "https://api.github.com",
+    accounts: [],
+    notes:
+      "GitHub integration for repository status, pull requests, issues, and future IDE/operator workflows. Add one entry per repository you want ZED to access.",
+  },
+  email: {
+    enabled: false,
+    status: "planned",
+    accounts: [],
+    notes:
+      "Outbound email lanes. Add one account per sending identity (e.g. an iCloud custom-domain alias, a Gmail SMTP credential, etc.). Each account holds its own SMTP host, port, username, and app password.",
+  },
+  google: {
+    enabled: false,
+    status: "planned",
+    accounts: [],
+    notes:
+      "Google account integration for Gmail, Calendar, and Drive access. Add one account per Google identity. Each account needs an OAuth client (from Google Cloud Console) plus a refresh token obtained via the initial consent flow.",
+  },
+  telephony: {
+    enabled: false,
+    status: "planned",
+    provider: "twilio",
+    phoneNumber: "",
+    voicemailEmail: "",
+    voiceAgentEnabled: false,
+    accountSid: "",
+    apiKey: "",
+    hasApiKey: false,
+    notes:
+      "Phone and voicemail lane for call routing, missed-call capture, voicemail summaries, and future voice assistant workflows.",
+  },
+  firewall: {
+    enabled: false,
+    status: "planned",
+    publicBaseUrl: "",
+    vpnBaseUrl: "",
+    preferredRoute: "vpn",
+    vpnProvider: "Tailscale",
+    authToken: "",
+    hasAuthToken: false,
+    healthPath: "/api/integration/firewall/status",
+    publicHealthPath: "/api/firewall/public-status",
+    zedAiWebhookBaseUrl: "",
+    notes:
+      "Fantasma Firewall integration for private VPN polling, custom-domain operator access, and future security workflow handoffs into ZED.",
+  },
+  businessOperations: {
+    enabled: true,
+    status: "configured",
+    ecommerce: true,
+    dropshipping: true,
+    realEstate: true,
+    acquisitions: true,
+    businessCredit: true,
+    rdSuggestions: true,
+    notes:
+      "Business Manager knowledge coverage for operations, acquisition planning, commerce, property, and business credit strategy.",
+  },
+  kalshi: {
+    enabled: false,
+    status: "planned",
+    environment: "demo",
+    apiBaseUrl: "https://demo-api.kalshi.co",
+    email: "",
+    notes:
+      "Planned event-market integration for market discovery, prediction research, and contract monitoring.",
+  },
+  voiceTranscription: {
+    enabled: false,
+    status: "browser-only",
+    provider: "Browser Speech API",
+  },
+  custom: [],
+};
